+++ lib/wavefront-sdk/credentials.rb
@@ -31,11 +31,16 @@ module Wavefront
     # @param options [Hash] keys may be 'file', which
     #   specifies a config file which will be loaded and parsed. If
     #   no file is supplied, those listed above will be used.;
-    #   and/or 'profile' which select a profile section from 'file'
+    #   and/or 'profile' which select a profile section from 'file'.
+    #   Specify the key :raise_noprofile to have an exception thrown
+    #   if a given profile cannot be found. Otherwise that is ignored and
+    #   options are built from other sources.
     #
     def initialize(options = {})
-      raw = load_from_file(cred_files(options), options[:profile] ||
-                           'default')
+      @raise_noprofile = options[:raise_on_no_profile] || false
+      raw = load_from_file(real_files(cred_files(options)),
+                           options[:profile] || 'default')
+
       populate(env_override(raw))
     end
 
@@ -80,31 +85,45 @@ module Wavefront
       end
     end
 
+    def real_files(files)
+      files.select { |f| f.exist? && f.file? }
+    end
+
     # @param files [Array][Pathname] a list of ini-style config files
     # @param profile [String] a profile name
+    # @param disallow_missing [Bool] whether or not to raise an exception if
+    #   we are given a profile but can't find it.
     # @return [Hash] the given profile from the given list of files.
     #   If multiple files match, the last one will be used
     #
     def load_from_file(files, profile = 'default')
       ret = {}
+      profiles_found = conf_files_found = 0
 
       files.each do |f|
-        next unless f.exist? && f.file?
-
         ret = load_profile(f, profile)
+        conf_files_found += 1
+        profiles_found += 1 unless ret.empty?
         ret[:file] = f
       end
 
+      raise_on_missing_profile(profile, profiles_found, conf_files_found)
       ret
     end
 
-    # Load in an (optionally) given section of an ini-style
-    # configuration file not there, we don't consider that an error.
+    def raise_on_missing_profile(profile, profiles_found, conf_files_found)
+      return true unless @raise_noprofile
+      return true unless profiles_found.zero? && conf_files_found.positive?
+
+      raise Wavefront::Exception::MissingConfigProfile, profile
+    end
+
+    # Load in an (optionally) given section of an ini-style configuration
+    # file. If the section is not there, we don't consider that an error.
     #
     # @param file [Pathname] the file to read
     # @param profile [String] the section in the config to read
-    # @return [Hash] options loaded from file. Each key becomes a
-    #   symbol
+    # @return [Hash] options loaded from file. Each key becomes a symbol
     #
     def load_profile(file, profile = 'default')
       IniFile.load(file)[profile].each_with_object({}) do |(k, v), memo|